# docs/architecture-pack/contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: Revizr API
  version: "1.0.0"
  description: |
    Revizr backend API — exam revision platform for UK students (KS3, GCSE, A-Level, 11+).
    All endpoints require Bearer JWT authentication unless noted as public.
    Base URL: https://api.revizr.co.uk/v1
  contact:
    name: Revizr Engineering
    email: engineering@revizr.co.uk
  license:
    name: Proprietary
servers:
  - url: https://api.revizr.co.uk/v1
    description: Production (eu-west-2)
  - url: https://api.staging.revizr.co.uk/v1
    description: Staging (eu-west-2)
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Access token issued by POST /auth/login or POST /auth/refresh.
        Expires in 15 minutes. Pass in Authorization header as "Bearer <token>".
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "email is required"
        details:
          type: object
          example: { "field": "email", "constraint": "format" }
...
    UserProfile:
      type: object
      required: [id, email, display_name, role, age_band, locale, created_at]
      properties:
        id:
          type: string
          format: uuid
          example: "8f274e3a-1c2d-4b5e-9f6a-7d8c9b0a1e2f"
        email:
          type: string
          format: email
          example: "amara@example.com"
        display_name:
          type: string
          example: "Amara"
        role:
          type: string
          enum: [student, parent, teacher]
          example: "student"
        age_band:
          type: string
          enum: [under13, 13to15, 16to18, adult]
          example: "13to15"
        locale:
          type: string
          example: "en-GB"
        created_at:
          type: string
          format: date-time
          example: "2026-05-01T09:00:00Z"
    Subject:
      type: object
      required: [id, user_id, subject_name, exam_board, level, is_active]
      properties:
        id:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        user_id:
          type: string
          format: uuid
          example: "8f274e3a-1c2d-4b5e-9f6a-7d8c9b0a1e2f"
        subject_name:
          type: string
          example: "Mathematics"
        exam_board:
          type: string
          example: "AQA"
        level:
          type: string
          enum: [11plus, ks3, gcse, alevel]
          example: "gcse"
        is_active:
          type: boolean
          example: true
...
  responses:
    Unauthorized:
      description: Missing or invalid access token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: "UNAUTHORIZED"
            message: "Invalid or expired access token"
...
    UnprocessableEntity:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: "VALIDATION_ERROR"
            message: "Request body failed validation"
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: "RATE_LIMITED"
            message: "Too many requests. Retry after 60 seconds."
...
paths:
...
  /users/me:
    get:
      tags: [users]
      operationId: getUsersMe
      summary: Get current user's profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      tags: [users]
      operationId: patchUsersMe
      summary: Update current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name:
                  type: string
                  maxLength: 100
                  example: "Amara K."
                locale:
                  type: string
                  enum: [en-GB, cy]
                  example: "cy"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
    delete:
      tags: [users]
      operationId: deleteUsersMe
      summary: Request GDPR account deletion
      description: >
        Initiates GDPR right-to-erasure. Soft-deletes the account and queues
        hard deletion of personal data fields within 30 days. Subscription is
        cancelled in Stripe immediately. Requires confirmation body to prevent
        accidental deletion.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [confirm]
              properties:
                confirm:
                  type: boolean
                  enum: [true]
                  example: true
      responses:
        "202":
          description: Deletion queued; account deactivated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Account deletion queued. Data will be erased within 30 days."
        "401":
          $ref: "#/components/responses/Unauthorized"
  /users/me/export:
    get:
      tags: [users]
      operationId: getUsersMeExport
      summary: GDPR data export — download all personal data
      description: >
        Returns a JSON document containing all personal data held for the current
        user across all entities. Scoped to C3+ fields the user is entitled to
        see. Rate-limited to 1 request per 24 hours per user.
      responses:
        "200":
          description: GDPR data export
          content:
            application/json:
              schema:
                type: object
                properties:
                  export_generated_at:
                    type: string
                    format: date-time
                    example: "2026-05-15T12:00:00Z"
                  user:
                    $ref: "#/components/schemas/UserProfile"
                  subjects:
                    type: array
                    items:
                      $ref: "#/components/schemas/Subject"
                  diagnostic_sessions:
                    type: array
                    items:
                      type: object
                      example: { "id": "b7c6d5e4-...", "session_type": "report_upload", "status": "complete" }
                  practice_sessions:
                    type: array
                    items:
                      type: object
                      example: { "id": "d4e5f6a7-...", "status": "completed", "question_count": 10 }
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ─── SUBJECTS ────────────────────────────────────────────────────────────────
...